--- i18n/locales/en/global/translations.ts
@@ -1,12 +1,0 @@
-export const translations = {
-  global: {
-    en: {
-      next: 'Continue',
-      copy: 'Copy',
-      chapter: 'Chapter',
-      challenge: 'Challenge',
-      coming_soon: 'Coming soon',
-      start_chapter: 'Start chapter',
-    },
-  },
-}
