+++ i18n/locales/en/index.ts
@@ -1,0 +1,3 @@
+import { translations as enTranslations } from './translations'
+
+export const translations = [enTranslations]
