--- i18n/locales/en/ui/translations.ts
@@ -1,96 +1,0 @@
-export const translations = {
-  end: {
-    en: {
-      save: 'Save my progress',
-      next: 'Continue without saving',
-    },
-  },
-  challenge_list: {
-    en: {
-      coming_soon: 'Coming soon. Hang tight.',
-    },
-  },
-  chapter: {
-    en: {
-      chapter_locked_one: 'Complete Chapter',
-      chapter_locked_two: 'to unlock.',
-    },
-  },
-  hero: {
-    en: {
-      title: 'Saving Satoshi',
-      description: 'Code your way through the mysteries of bitcoin.',
-      start_journey: 'Start your journey',
-      tell_more: 'Tell me more',
-    },
-
-    nl: {
-      title: 'Saving Satoshi',
-      description: 'Programmeer je weg door de mysteries van bitcoin.',
-      start_journey: 'Start je reis',
-      tell_more: 'Vertel me meer',
-    },
-  },
-
-  footer: {
-    en: {
-      paragraph_one: 'An open-source production by the bitcoin community.',
-      link: 'Check the code',
-    },
-  },
-
-  navbar: {
-    en: {
-      intro: 'Intro',
-      chapter: 'Chapter',
-      chapter_complete: 'Chapter complete',
-      challenge: 'Challenge',
-    },
-
-    nl: {
-      intro: 'Intro',
-      chapter: 'Hoofdstuk',
-      chapter_complete: 'Hoofdstuk compleet',
-      challenge: 'Uitdaging',
-    },
-  },
-
-  modal_login: {
-    en: {
-      heading: 'Load your progress',
-      paragraph_one:
-        'Enter your personal code below to restore your chapter and challenge progress, as well as your avatar.',
-      clear: 'No code yet?',
-      prompt: 'Enter your code',
-      confirm: 'Load my progress',
-      paragraph_two:
-        'Sorry, that’s not the right code for the progress stored in this browser.',
-      paragraph_three: 'No code yet or want to start again?',
-      quit: 'Clear save progress.',
-      success: 'Progress loaded',
-      success_message: 'You’re all set to continue with Chapter',
-    },
-  },
-
-  modal_logout: {
-    en: {
-      heading: 'You are logged in',
-      paragraph_one:
-        'Your progress is stored in this browser. You will be automatically logged out after 30 days. Use the code below to restore it after taking a break.',
-      signout: 'Sign out',
-    },
-  },
-
-  modal_signup: {
-    en: {
-      heading: 'Want to save your progress?',
-      paragraph_one:
-        'Copy and store a simple code to save and load your progress on any device and browser. If you already have a code, load your progress here.',
-      subheading_one: 'Choose an avatar',
-      subheading_two: 'Back up your personal code',
-      paragraph_two:
-        'All set? Code copied and backed-up? Make sure your do, as it can’t be recovered if you lose it.',
-      confirm: 'I’ve copied and backed up my code',
-    },
-  },
-}
