--- i18n/locales/nl/content/translations.ts
@@ -1,47 +1,0 @@
-export const translations = {
-  chapter_one: {
-    title: 'Verborgen geheimen in het volle zicht',
-    paragraph_one:
-      'Het is het jaar 2139. Over twee weken zal de allerlaatste bitcoin worden gemined. Al maanden wordt er op Times Square afgeteld voor het laatste block waarna er officieel geen bitcoin meer zal worden uitgegeven.',
-    paragraph_two: `En plotseling komt het netwerk tot stilstand. In plaats van het normale blockinterval van tien minuten, komen er maar een paar keer per dag blocks binnen. De bitcoin-node van je moeder stopt met synchroniseren. De media beginnen verslag uit te brengen over de situatie. Het is overal op sociale media.`,
-    paragraph_three: `De prijs van bitcoin daalt. Er wordt op je deur geklopt en als je opendoet, ligt er een vergeelde, verweerde envelop voor je klaar. Er is geen naam of retouradres. Wat zou erin kunnen zitten?`,
-  },
-  chapter_two: {
-    title: 'De 51% aanval',
-    paragraph_one:
-      'De netwerkstoring moet iets te maken hebben met de block subsidie. Dit kan geen toeval zijn',
-    paragraph_two:
-      'Je springt achter je computer om wat onderzoek te doen naar minen wanneer je iets ongewoons opmerkt aan de weinige blocks die worden gemined. Ze lijken allemaal op dezelfde plaats te worden gemined, mogelijk zelfs uit dezelfde mining pool. Wat is er met de rest van de miners gebeurd?',
-  },
-  chapter_three: {
-    title: 'Je beloning claimen',
-  },
-  chapter_four: {
-    title: 'Wil de echte Satoshi opstaan?',
-  },
-  chapter_five: {
-    title: 'De sleutelhouder',
-  },
-  chapter_six: {
-    title: 'Aanvallen is de beste verdediging',
-  },
-  chapter_seven: {
-    title: 'Twentyone',
-  },
-  chapter_eight: {
-    title: 'Een kwestie van duurzaamheid',
-  },
-  chapter_nine: {
-    title: `Niet berekenen, verifiëren.`,
-  },
-  chapter_ten: {
-    title: '10 miljard verbindingen',
-  },
-  reaping_rewards: {
-    title: 'Blocks en miners',
-    subtitle:
-      '"Om het mysterie en de magie te begrijpen, moet je bij het begin beginnen…"',
-    paragraph_one: `Satoshi, de pseudonieme uitvinder van bitcoin, ontgonnen ook het eerste bitcoin-blok. Hij liet je een verborgen aanwijzing achter in de allereerste bitcoin-transactie die ooit is gemaakt.`,
-    paragraph_two: `Elk blok in de bitcoin-blockchain bevat een reeks transacties. Een leuke manier om deze blokken en de dingen die ze bevatten te bekijken, is door een blokverkenner te gebruiken.`,
-  },
-}
