--- i18n/locales/nl/global/translations.ts
@@ -1,3 +1,0 @@
-export const translations = {
-  global: {},
-}
