+++ i18n/locales/nl/index.ts
@@ -1,0 +1,3 @@
+import { translations as nlTranslations } from './translations'
+
+export const translations = [nlTranslations]
