+++ i18n/locales/nl/translations.ts
@@ -1,0 +1,116 @@
+export const translations = {
+  chapter_one: {
+    nl: {
+      title: 'Verborgen geheimen in het volle zicht',
+      paragraph_one:
+        'Het is het jaar 2139. Over twee weken zal de allerlaatste bitcoin worden gemined. Al maanden wordt er op Times Square afgeteld voor het laatste block waarna er officieel geen bitcoin meer zal worden uitgegeven.',
+      paragraph_two: `En plotseling komt het netwerk tot stilstand. In plaats van het normale blockinterval van tien minuten, komen er maar een paar keer per dag blocks binnen. De bitcoin-node van je moeder stopt met synchroniseren. De media beginnen verslag uit te brengen over de situatie. Het is overal op sociale media.`,
+      paragraph_three: `De prijs van bitcoin daalt. Er wordt op je deur geklopt en als je opendoet, ligt er een vergeelde, verweerde envelop voor je klaar. Er is geen naam of retouradres. Wat zou erin kunnen zitten?`,
+    },
+  },
+  chapter_two: {
+    nl: {
+      title: 'De 51% aanval',
+      paragraph_one:
+        'De netwerkstoring moet iets te maken hebben met de block subsidie. Dit kan geen toeval zijn',
+      paragraph_two:
+        'Je springt achter je computer om wat onderzoek te doen naar minen wanneer je iets ongewoons opmerkt aan de weinige blocks die worden gemined. Ze lijken allemaal op dezelfde plaats te worden gemined, mogelijk zelfs uit dezelfde mining pool. Wat is er met de rest van de miners gebeurd?',
+    },
+  },
+  chapter_three: {
+    nl: {
+      title: 'Je beloning claimen',
+    },
+  },
+  chapter_four: {
+    nl: {
+      title: 'Wil de echte Satoshi opstaan?',
+    },
+  },
+  chapter_five: {
+    nl: {
+      title: 'De sleutelhouder',
+    },
+  },
+  chapter_six: {
+    nl: {
+      title: 'Aanvallen is de beste verdediging',
+    },
+  },
+  chapter_seven: {
+    nl: {
+      title: 'Twentyone',
+    },
+  },
+  chapter_eight: {
+    nl: {
+      title: 'Een kwestie van duurzaamheid',
+    },
+  },
+  chapter_nine: {
+    nl: {
+      title: `Niet berekenen, verifiëren.`,
+    },
+  },
+  chapter_tnl: {
+    nl: {
+      title: '10 miljard verbindingen',
+    },
+  },
+  reaping_rewards: {
+    nl: {
+      title: 'Blocks en miners',
+      subtitle:
+        '"Om het mysterie en de magie te begrijpen, moet je bij het begin beginnen…"',
+      paragraph_one: `Satoshi, de pseudonieme uitvinder van bitcoin, ontgonnen ook het eerste bitcoin-blok. Hij liet je een verborgen aanwijzing achter in de allereerste bitcoin-transactie die ooit is gemaakt.`,
+      paragraph_two: `Elk blok in de bitcoin-blockchain bevat een reeks transacties. Een leuke manier om deze blokken en de dingen die ze bevatten te bekijken, is door een blokverkenner te gebruiken.`,
+    },
+  },
+  genesis_one: {
+    nl: {
+      title: 'Genesis',
+      heading: 'Vind block 0',
+      paragraph_one: `Laten we het allereerste block in de Bitcoin blockchain vinden. Gebruik een <Tooltip content="genesis_one.tooltip_block_explorer" href="https://blockstream.info" className="underline">block explorer</Tooltip> en zoek naar ’0’. Dit zal je naar het block op hoogte 0 nemen. Zoals je kunt zien bevat dit blok maar één transactie.`,
+      paragraph_two: `Klap de transactie uit zodat u de details beter kunt bekijken. Aan de linkerzijde is er een input genaamd 'Coinbase'. Daarbinnen bevindt zich een veld met de naam 'ScriptSig'. Kopieer de <Tooltip content="genesis_one.tooltip_hex" className="underline">HEX</Tooltip> versie van de ScriptSig en plak deze hieronder.`,
+      tooltip_block_explorer: 'We recommend https://blockstream.info/',
+      tooltip_hex:
+        'Short for hexadecimal, a numeral system that represents numbers using a base of 16.',
+      view_block_0: 'Vind Block 0',
+      placeholder: 'Plak de HEX versie van de ScriptSig',
+    },
+  },
+  genesis_two: {
+    nl: {
+      title: 'Genesis',
+      heading: 'Laten we dit duidelijk maken',
+      waiting_for_input: 'Wachten tot de script wordt uitgevoerd...',
+      success:
+        'Interessant! Het bericht verwijst naar een nieuwsartikel dat op dezelfde dag dat Satoshi het genesis block heeft gemined, gepubliceerd is. Dit geeft ons een belangrijke aanwijzing over de motivatie van Satoshi voor het maken van Bitcoin. Je kunt er <Link href="https://bitcoinmagazine.com/culture/ten-years-later-reflection-bitcoins-genesis-and-satoshis-timing" className="underline">hier</Link> meer over lezen.',
+    },
+  },
+  transacting_one: {
+    nl: {},
+  },
+  transacting_two: {
+    nl: {},
+  },
+  transacting_three: {
+    nl: {},
+  },
+  hero: {
+    nl: {
+      title: 'Saving Satoshi',
+      description: 'Programmeer je weg door de mysteries van bitcoin.',
+      start_journey: 'Start je reis',
+      tell_more: 'Vertel me meer',
+    },
+  },
+  navbar: {
+    nl: {
+      intro: 'Intro',
+      chapter: 'Hoofdstuk',
+      chapter_complete: 'Hoofdstuk compleet',
+      challenge: 'Uitdaging',
+    },
+  },
+}
