+++ middleware.ts
@@ -5,7 +5,7 @@ import { NextRequest, NextResponse } from 'next/server'
 
 import { match as matchLocale } from '@formatjs/intl-localematcher'
 import Negotiator from 'negotiator'
-import { i18n } from 'config/i18n'
+import { i18n } from 'i18n/config'
 
 function getLocale(request: NextRequest): string | undefined {
   // Negotiator expects plain object so we need to transform headers
